Editable actions fix
