+++ src/demo.js
@@ -19,7 +19,7 @@ class App extends Component {
   state = {
     selecteds: 0,
     data: [
-      { id: 1, name: 'axxxxxasdasdasdasd', surname: 'Baran', isMarried: true, birthDate: new Date(1987, 1, 1), birthCity: 0, sex: 'Male', type: 'adult', insertDateTime: new Date(2018, 1, 1, 12, 23, 44), time: new Date(1900, 1, 1, 14, 23, 35) },
+      { id: 1, name: 'axxxxxasdasdasdasd', surname: 'Baran', isMarried: true, birthDate: new Date(1987, 1, 1), birthCity: 34, sex: 'Male', type: 'adult', insertDateTime: new Date(2018, 1, 1, 12, 23, 44), time: new Date(1900, 1, 1, 14, 23, 35) },
       { id: 2, name: 'bxxxxxasdasdasdasd', surname: 'Baran', isMarried: false, birthDate: new Date(1987, 1, 1), birthCity: 34, sex: 'Female', type: 'adult', insertDateTime: new Date(2018, 1, 1, 12, 23, 44), time: new Date(1900, 1, 1, 14, 23, 35), parentId: 1 },
       { id: 3, name: 'cxxxxxasdasdasdasd', surname: 'Baran', isMarried: true, birthDate: new Date(1987, 1, 1), birthCity: 34, sex: 'Female', type: 'child', insertDateTime: new Date(2018, 1, 1, 12, 23, 44), time: new Date(1900, 1, 1, 14, 23, 35), parentId: 1 },
       { id: 4, name: 'dxxxxxasdasdasdasd', surname: 'Baran', isMarried: true, birthDate: new Date(1987, 1, 1), birthCity: 34, sex: 'Female', type: 'child', insertDateTime: new Date(2018, 1, 1, 12, 23, 44), time: new Date(1900, 1, 1, 14, 23, 35), parentId: 3 },
@@ -59,11 +59,35 @@ class App extends Component {
                 ref={this.tableRef}
                 columns={this.state.columns}
                 data={this.state.data}
-                title="Demo Title"
-                parentChildData={(row, rows) => rows.find(a => a.id === row.parentId)}
-                options={{
-                  selection: true,
-                  defaultExpanded: true
+                title="Demo Title"                
+                editable={{
+                  onRowAdd: (newData) => new Promise((resolve, reject) => {
+                    setTimeout(() => {
+                      {/* const data = this.state.data;
+                      data.push(newData);
+                      this.setState({ data }, () => resolve()); */}
+            
+                      resolve();
+                    }, 1000);
+                  }),
+                  onRowUpdate: (newData, oldData) => new Promise((resolve, reject) => {
+                    setTimeout(() => {
+                      {/* const data = this.state.data;
+                      const index = data.indexOf(oldData);
+                      data[index] = newData;                
+                      this.setState({ data }, () => resolve()); */}
+                      resolve();
+                    }, 1000);
+                  }),
+                  onRowDelete: (oldData) => new Promise((resolve, reject) => {
+                    setTimeout(() => {
+                      {/* let data = this.state.data;
+                      const index = data.indexOf(oldData);
+                      data.splice(index, 1);
+                      this.setState({ data }, () => resolve()); */}
+                      resolve();
+                    }, 1000);
+                  }),
                 }}
               />
             </Grid>
